backend/requirements.txt
Flask
bs4
requests
openai
python-dotenv